Pull resolution proofing into a new module
