+++ app/jobs/resolution_proofing_job.rb
@@ -32,6 +32,11 @@ class ResolutionProofingJob < ApplicationJob
 
     applicant_pii = decrypted_args[:applicant_pii]
 
+    @resolution_proofer = Proofing::Resolution::Proofer.new(
+      should_proof_state_id: should_proof_state_id,
+      capture_secondary_id_enabled: capture_secondary_id_enabled,
+    )
+
     user = User.find_by(id: user_id)
 
     optional_threatmetrix_result = proof_lexisnexis_ddp_with_threatmetrix_if_needed(
@@ -70,6 +75,8 @@ class ResolutionProofingJob < ApplicationJob
   end
 
   private
+
+  attr_reader :resolution_proofer
 
   def log_threatmetrix_info(threatmetrix_result, user)
     logger_info_hash(
@@ -128,39 +135,13 @@ class ResolutionProofingJob < ApplicationJob
   end
 
   # @return [CallbackLogData]
-  def proof_lexisnexis_then_aamva(timer:, applicant_pii:, should_proof_state_id:,
-                                  capture_secondary_id_enabled:)
-    # Proof the user's state ID address if the user is going through the double address
-    # verification flow and has a different address listed on their state ID. Otherwise
-    # proof their residential address
-    if capture_secondary_id_enabled && applicant_pii[:same_address_as_id] == 'false'
-      applicant_pii = with_state_id_address(applicant_pii)
-    end
-
-    resolution_result = timer.time('resolution') do
-      # todo (LG-8693): Proof the user's state ID address in addition to (rather than instead
-      # of) their residential address
-      resolution_proofer.proof(applicant_pii)
-    end
-
-    state_id_result = Proofing::StateIdResult.new(
-      success: true, errors: {}, exception: nil, vendor_name: 'UnsupportedJurisdiction',
-    )
-    if should_proof_state_id && user_can_pass_after_state_id_check?(resolution_result)
-      timer.time('state_id') do
-        state_id_result = state_id_proofer.proof(applicant_pii)
-      end
-    end
-
-    result = Proofing::ResolutionResultAdjudicator.new(
-      resolution_result: resolution_result,
-      state_id_result: state_id_result,
-      should_proof_state_id: should_proof_state_id,
-    ).adjudicated_result.to_h
+  def proof_lexisnexis_then_aamva(timer:)
+    result = resolution_proofer.proof(applicant_pii: @applicant_pii, timer: timer)
 
     CallbackLogData.new(
-      result: result,
+      residential_address_success: residential_address_result.success?,
       resolution_success: resolution_result.success?,
+      result: result.adjudicated_result.to_h,
       state_id_success: state_id_result.success?,
     )
   end
@@ -179,22 +160,6 @@ class ResolutionProofingJob < ApplicationJob
     results_that_cannot_pass_aamva.blank?
   end
 
-  def resolution_proofer
-    @resolution_proofer ||=
-      if IdentityConfig.store.proofer_mock_fallback
-        Proofing::Mock::ResolutionMockClient.new
-      else
-        Proofing::LexisNexis::InstantVerify::Proofer.new(
-          instant_verify_workflow: IdentityConfig.store.lexisnexis_instant_verify_workflow,
-          account_id: IdentityConfig.store.lexisnexis_account_id,
-          base_url: IdentityConfig.store.lexisnexis_base_url,
-          username: IdentityConfig.store.lexisnexis_username,
-          password: IdentityConfig.store.lexisnexis_password,
-          request_mode: IdentityConfig.store.lexisnexis_request_mode,
-        )
-      end
-  end
-
   def lexisnexis_ddp_proofer
     @lexisnexis_ddp_proofer ||=
       if IdentityConfig.store.lexisnexis_threatmetrix_mock_enabled
@@ -208,46 +173,10 @@ class ResolutionProofingJob < ApplicationJob
       end
   end
 
-  def state_id_proofer
-    @state_id_proofer ||=
-      if IdentityConfig.store.proofer_mock_fallback
-        Proofing::Mock::StateIdMockClient.new
-      else
-        Proofing::Aamva::Proofer.new(
-          auth_request_timeout: IdentityConfig.store.aamva_auth_request_timeout,
-          auth_url: IdentityConfig.store.aamva_auth_url,
-          cert_enabled: IdentityConfig.store.aamva_cert_enabled,
-          private_key: IdentityConfig.store.aamva_private_key,
-          public_key: IdentityConfig.store.aamva_public_key,
-          verification_request_timeout: IdentityConfig.store.aamva_verification_request_timeout,
-          verification_url: IdentityConfig.store.aamva_verification_url,
-        )
-      end
-  end
-
   def add_threatmetrix_proofing_component(user_id, threatmetrix_result)
     ProofingComponent.
       create_or_find_by(user_id: user_id).
       update(threatmetrix: true,
              threatmetrix_review_status: threatmetrix_result.review_status)
   end
-
-  # Make a copy of pii with the user's state ID address overwriting the address keys
-  def with_state_id_address(pii)
-    pii_with_state_id_address = pii.transform_keys(SECONDARY_ID_ADDRESS_MAP)
-    # todo (LG-9237): Remove the below lines once the user's state ID address state
-    # is saved as :state_id_state
-    pii_with_state_id_address[:state_id_jurisdiction] = pii[:state_id_jurisdiction]
-    pii_with_state_id_address
-  end
-
-  SECONDARY_ID_ADDRESS_MAP = {
-    state_id_address1: :address1,
-    state_id_address2: :address2,
-    state_id_city: :city,
-    # todo (LG-9237): Change below to `state_id_state: :state` once we are collecting
-    # user's state ID address state as :state_id_state
-    state_id_jurisdiction: :state,
-    state_id_zipcode: :zipcode,
-  }.freeze
 end
